Commenting

+++ app/Http/Controllers/EmailController.php
@@ -8,6 +8,8 @@ use Illuminate\Support\Facades\Mail;
 use Symfony\Component\Mailer\Exception\TransportException;
 use App\Models\Account;
 use App\Models\Message;
+use App\Models\EmailPreference;
+use DateTime;
 
 class EmailController extends Controller
 {
@@ -132,11 +134,14 @@ class EmailController extends Controller
         // Mail::to("{$accountNo}@curtin.edu.au")->send(new MJML("Nomination Reminder", "email/nominationReminder", $dynamicData));
     }
 
+
+    // Goes through the backlog of unset emails and attempts to send them
     public function attemptBacklog(): void {
-        //TODO add function to check email backlog
         $backlog = UnsentEmail::all();
         foreach($backlog as $email)
         {
+            // switch case used here for expandability to other email types
+            // that might be added later
             $subject = $email->subject;
             switch($subject){
                 case "Unacknowledged Messages":
@@ -146,17 +151,34 @@ class EmailController extends Controller
         }
     }
 
+
+    // Handles the attempted resending of an archive email
     private function attemptUnackMsg($email)
     {
         try
         {
+            // Get account and preferences
             $accountNo = $email->accountNo;
             $user = Account::where('accountNo', $accountNo)->first();
-            $messages = Message::where('receiverNo', $user->accountNo)->where('acknowledged', 0)->get();
-            if ($messages->count() != 0) { // if Has messages
-                $user->sendDailyMessageNotification($messages);
+            $preferences = EmailPreference::get()->where('accountNo', $accountNo)->first();
+
+            // calculate hours between current time and last time the reminder was sent
+            $now = new DateTime('NOW');
+            $lastSent = new DateTime($preferences->timeLastSent);
+            $interval = $now->diff($lastSent);
+            $hourInterval = ($interval->h) + ($interval->days * 24);
+
+            // if the time passed is greater than the reminder frequency
+            $frequency = $preferences->hours;
+            if( $hourInterval > $frequency)
+            {
+                // try and send the email if they have any messages
+                $messages = Message::where('receiverNo', $user->accountNo)->where('acknowledged', 0)->get();
+                if ($messages->count() != 0) { // if Has messages
+                    $user->sendDailyMessageNotification($messages);
+                }
+                $email->delete(); // delete from backlog
             }
-            $email->delete();
         }
         catch(TransportException $e)
         {
